Replace Geocoder::Configuration.test_mode with implementation on Geocoder::Lookup::Test
