+++ lib/geocoder/lookups/test.rb
@@ -5,10 +5,26 @@ module Geocoder
   module Lookup
     class Test < Base
 
+      def self.add_stub(query, results)
+        stubs[query] = results
+      end
+
+      def self.read_stub(query)
+        stubs.fetch(query) { raise ArgumentError, "unknown stub request #{query}" }
+      end
+
+      def self.stubs
+        @stubs ||= {}
+      end
+
+      def self.reset
+        @stubs = {}
+      end
+
       private
 
       def results(query, reverse = false)
-        Geocoder::Configuration.read_stub(query)
+        Geocoder::Lookup::Test.read_stub(query)
       end
 
     end
